changed y from int16_t to int32_t to prevent overflow

+++ EcoRob/boost_pwm.ino
@@ -55,7 +55,7 @@ ISR(TIMER0_COMPA_vect)
   static uint8_t isr_cntr = 0;
   int16_t xd = 0;
   static int16_t xdsum = 0;
-  static int16_t y = 0;
+  static int32_t y = 0;
   
   
   if(isr_cntr < ADC_MAX_CH)
@@ -66,7 +66,7 @@ ISR(TIMER0_COMPA_vect)
     xd = BOOST_OUTPUT_REF - BOOST_U_OUT;
 
     xdsum += xd;
-    y = (uint32_t)KP * xd + KI * xdsum;
+    y = (int32_t)KP * xd + KI * xdsum;
 
     // anti-windup
     // shift left (11.5 fixed point)
